Move import/export code to its own file
