+++ CodeLite/pwsafe-cli.project
@@ -42,6 +42,8 @@
     <File Name="../src/ui/cli/diff.h"/>
     <File Name="../src/ui/cli/safeutils.cpp"/>
     <File Name="../src/ui/cli/safeutils.h"/>
+    <File Name="../src/ui/cli/impexp.cpp"/>
+    <File Name="../src/ui/cli/impexp.h"/>
   </VirtualDirectory>
   <Dependencies Name="Debug">
     <Project Name="core"/>
