+++ src/ui/cli/impexp.h
@@ -1,0 +1,8 @@
+#pragma once
+
+class PWScore;
+class UserArgs;
+
+int Import(PWScore &core, const UserArgs &ua);
+int Export(PWScore &core, const UserArgs &ua);
+
